utilisation de intptr_t au lieu de long long dans ConstAssignClause

+++ ConstAssignClause.hh
@@ -2,6 +2,8 @@
 #ifndef CONSTASSIGNCLAUSE_HH
 #define CONSTASSIGNCLAUSE_HH
 
+#include <stdint.h> // pour intptr_t
+
 class ConstAssignClause HERITED_CLAUSE
 {
 public:
@@ -23,7 +25,7 @@ public:
     const unsigned _number;
     #endif
 protected:
-    long long _currentHash; // sum of the adresses of the FREE Variables contained in the ConstAssignClause
+    intptr_t _currentHash; // sum of the adresses of the FREE Variables contained in the ConstAssignClause
     bool _currentHashVal; // xor of all the pos() of the free literals contained in the ConstAssignClause
     Variable* _satisfied; // the adresse of the first TRUE Literal of the ConstAssignClause
     unsigned int _numOfFree; // number of FREE Variables contained in the ConstAssignClause
