+++ ConstAssignClause.inline.hh
@@ -8,12 +8,12 @@
 
 
 inline ConstAssignClause::ConstAssignClause(CONSTR_ARGS(list))
-    : INIT_FOR_VERBOSE()  _currentHash((long long)NULL), _currentHashVal(false), _satisfied(NULL), _numOfFree(list.size())
+    : INIT_FOR_VERBOSE()  _currentHash((intptr_t)NULL), _currentHashVal(false), _satisfied(NULL), _numOfFree(list.size())
 {
     std::vector<Literal>::const_iterator it;
     for(it = list.begin(); it != list.end(); ++it)
     {
-        _currentHash += (long long)it->var();
+        _currentHash += (intptr_t)it->var();
         _currentHashVal = (_currentHashVal != it->pos()); // XOR booléen
         it->var()->linkToClause(it->pos(), (StockedClause*)this);
     }
@@ -24,7 +24,7 @@ inline bool ConstAssignClause::setLitFalse(const Literal& l)
 {
     if(_satisfied == NULL)
     {
-        _currentHash -= (long long)l.var();
+        _currentHash -= (intptr_t)l.var();
         _currentHashVal = (_currentHashVal != !l.pos()); // XOR booléen
         _numOfFree--;
     }
@@ -46,7 +46,7 @@ inline void ConstAssignClause::freeLitFalse(const Literal& l)
 {
     if(_satisfied == NULL)
     {
-        _currentHash += (long long)l.var();
+        _currentHash += (intptr_t)l.var();
         _currentHashVal = (_currentHashVal != !l.pos()); // XOR booléen
         _numOfFree++;
     }
